Timer1_Delay

+++ MCAL/04-Timer1/TIMER1_interface.h
@@ -4,8 +4,7 @@
 
 void Timer1_voidInit(void);
 void Timer1_voidEnableOVERFLOW_INTERRUPT(void);
-void Timer1_voidDelay_Micro_Seconds_Sync(u16 Copy_u16Delay_micro_Seconds);
-//void Timer1_voidDelay_Micro_Seconds_Async(u16 Copy_u16Delay_micro_Seconds, void *ptr_ISR(void));
+void Timer1_voidDelay_Micro_Seconds_Sync(u32 Copy_u16Delay_micro_Seconds);
